+++ custom_components/supernotify/notify.py
@@ -85,6 +85,7 @@ from .transports.ntfy import NtfyTransport
 from .transports.persistent import PersistentTransport
 from .transports.pushover import PushoverTransport
 from .transports.sms import SMSTransport
+from .transports.smtp import SmtpTransport
 from .transports.telegram import TelegramTransport
 from .transports.tts import TTSTransport
 
@@ -119,6 +120,7 @@ TRANSPORTS: list[type[Transport]] = [
     TelegramTransport,
     LaMetricTransport,
     PushoverTransport,
+    SmtpTransport,
 ]  # No auto-discovery of transport plugins so manual class registration required here